(#3182) - avoid on-the-fly idb migrations
I was getting tired of the code complexity with
having to figure out whether `metadata.winningRev`
and `metadata.seq` were there or not. I prefer to
just migrate everything to be sure.

This also removes the docStore's 'seq' index, which
we weren't using and which apparently doesn't offer
any performance benefit.